# openapi/schema.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "summary": "KokoroNotoAPI_WithCI", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/entries": {"get": {"tags": ["entries"], "summary": "Get Entries", "operationId": "get_entries", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EntriesResponse"}}}}}}, "post": {"tags": ["entries"], "summary": "Add Entry", "operationId": "add_entry", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Entry-Input"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EntryResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["entries"], "summary": "Update Entry", "operationId": "update_entry", "parameters": [{"name": "id", "in": "query", "required": true, "schema": {"type": "string", "description": "\u30a8\u30f3\u30c8\u30ea\u30fcID", "title": "Id"}, "description": "\u30a8\u30f3\u30c8\u30ea\u30fcID"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Entry-Input"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EntryResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["entries"], "summary": "Delete Entry", "operationId": "delete_entry", "parameters": [{"name": "id", "in": "query", "required": true, "schema": {"type": "string", "description": "\u524a\u9664\u3059\u308b\u30a8\u30f3\u30c8\u30ea\u30fc\u306eID", "title": "Id"}, "description": "\u524a\u9664\u3059\u308b\u30a8\u30f3\u30c8\u30ea\u30fc\u306eID"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EntriesResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"EntriesResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "entries": {"items": {"$ref": "#/components/schemas/Entry-Output"}, "type": "array", "title": "Entries"}}, "type": "object", "required": ["status", "entries"], "title": "EntriesResponse"}, "Entry-Input": {"properties": {"_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Id", "description": "\u30a8\u30f3\u30c8\u30ea\u30fcID\uff08\u81ea\u52d5\u751f\u6210\u3001\u4efb\u610f\uff09", "example": "dummy_id"}, "record_date": {"type": "string", "format": "date", "title": "Record Date", "description": "\u8a18\u9332\u65e5\uff08\u5fc5\u9808\u3001YYYY-MM-DD\u5f62\u5f0f\uff09", "example": "2025-08-14"}, "mood_score": {"type": "integer", "maximum": 5.0, "minimum": 0.0, "title": "Mood Score", "description": "\u30e1\u30f3\u30bf\u30eb\u30b9\u30b3\u30a2\uff08\u5fc5\u9808\u30010\u301c5\u306e\u6574\u6570\uff09", "example": 4}, "sleep_hours": {"type": "number", "minimum": 0.0, "title": "Sleep Hours", "description": "\u7761\u7720\u6642\u9593\uff08\u5fc5\u9808\u30010\u4ee5\u4e0a\u306e\u5c0f\u6570\uff09", "example": 6.5}, "memo": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Memo", "description": "\u30e1\u30e2\uff08\u4efb\u610f\u3001\u7a7a\u6587\u5b57\u53ef\uff09", "example": "\u4eca\u65e5\u306f\u3088\u304f\u7720\u308c\u305f"}}, "additionalProperties": false, "type": "object", "required": ["record_date", "mood_score", "sleep_hours"], "title": "Entry"}, "Entry-Output": {"properties": {"id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Id", "description": "\u30a8\u30f3\u30c8\u30ea\u30fcID\uff08\u81ea\u52d5\u751f\u6210\u3001\u4efb\u610f\uff09", "example": "dummy_id"}, "record_date": {"type": "string", "title": "Record Date", "description": "\u8a18\u9332\u65e5\uff08\u5fc5\u9808\u3001YYYY-MM-DD\u5f62\u5f0f\uff09", "example": "2025-08-14"}, "mood_score": {"type": "integer", "maximum": 5.0, "minimum": 0.0, "title": "Mood Score", "description": "\u30e1\u30f3\u30bf\u30eb\u30b9\u30b3\u30a2\uff08\u5fc5\u9808\u30010\u301c5\u306e\u6574\u6570\uff09", "example": 4}, "sleep_hours": {"type": "number", "minimum": 0.0, "title": "Sleep Hours", "description": "\u7761\u7720\u6642\u9593\uff08\u5fc5\u9808\u30010\u4ee5\u4e0a\u306e\u5c0f\u6570\uff09", "example": 6.5}, "memo": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Memo", "description": "\u30e1\u30e2\uff08\u4efb\u610f\u3001\u7a7a\u6587\u5b57\u53ef\uff09", "example": "\u4eca\u65e5\u306f\u3088\u304f\u7720\u308c\u305f"}}, "additionalProperties": false, "type": "object", "required": ["record_date", "mood_score", "sleep_hours"], "title": "Entry"}, "EntryResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "entry": {"$ref": "#/components/schemas/Entry-Output"}}, "type": "object", "required": ["status", "entry"], "title": "EntryResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}, "tags": [{"name": "entries", "description": "\u30e1\u30f3\u30bf\u30eb\u30d8\u30eb\u30b9\u8a18\u9332\u30a8\u30f3\u30c8\u30ea\u30fc\u306e\u7ba1\u7406\u64cd\u4f5c\u3002\u65e5\u3005\u306e\u6c17\u5206\u30b9\u30b3\u30a2\u3001\u7761\u7720\u6642\u9593\u3001\u30e1\u30e2\u306a\u3069\u3092\u8a18\u9332\u30fb\u7ba1\u7406\u3067\u304d\u307e\u3059\u3002"}]}